+++ MezonQuiz/src/Mezon-sdk/Structrues/User.cs
@@ -9,7 +9,7 @@ namespace Mezon_sdk.Structures
     {
         public int Id { get; set; }
         public string? Avatar { get; set; }
-        public int? DmChannelId { get; set; }
+        public long? DmChannelId { get; set; }
         public string? Username { get; set; }
         public string? ClanNick { get; set; }
         public string? ClanAvatar { get; set; }
